Making Observer -> ObserverJob and extracting filters into ObserverFilters
