+++ Source/Kernel/Grains/Observation/CatchUp.cs
@@ -15,7 +15,7 @@ namespace Aksio.Cratis.Kernel.Grains.Observation;
 /// <summary>
 /// Represents an implementation of <see cref="ICatchUp"/>.
 /// </summary>
-public class CatchUp : Observer, ICatchUp
+public class CatchUp : ObserverJob, ICatchUp
 {
     readonly ILogger<CatchUp> _logger;
     ObserverKey? _observerKey;
