+++ Source/Kernel/Grains/Observation/ObserverLogMessages.cs
@@ -12,8 +12,8 @@ namespace Aksio.Cratis.Kernel.Grains.Observation;
 internal static partial class ObserverLogMessages
 {
     [LoggerMessage(0, LogLevel.Information, "Observer {ObserverId} is now active up for microservice '{MicroserviceId}' on sequence '{EventSequenceId}' for tenant '{TenantId}'")]
-    internal static partial void Active(this ILogger<Observer> logger, ObserverId observerId, MicroserviceId microserviceId, EventSequenceId eventSequenceId, TenantId tenantId);
+    internal static partial void Active(this ILogger<ObserverJob> logger, ObserverId observerId, MicroserviceId microserviceId, EventSequenceId eventSequenceId, TenantId tenantId);
 
     [LoggerMessage(1, LogLevel.Warning, "Partition {Partition} failed for event with sequence number {EventSequenceNumber} observer {ObserverId} for sequence {EventSequenceId} for microservice '{MicroserviceId}' and tenant '{TenantId}' - observing source microservice '{SourceMicroserviceId}' and tenant '{SourceTenantId}'")]
-    internal static partial void PartitionFailed(this ILogger<Observer> logger, string partition, EventSequenceNumber eventSequenceNumber, ObserverId observerId, EventSequenceId eventSequenceId, MicroserviceId microserviceId, TenantId tenantId, MicroserviceId sourceMicroserviceId, TenantId sourceTenantId);
+    internal static partial void PartitionFailed(this ILogger<ObserverJob> logger, string partition, EventSequenceNumber eventSequenceNumber, ObserverId observerId, EventSequenceId eventSequenceId, MicroserviceId microserviceId, TenantId tenantId, MicroserviceId sourceMicroserviceId, TenantId sourceTenantId);
 }
